fix(ci): repair the rustdoc gate, Windows fixture digests, and a dtype-dependent test
CI has been failing on main since the 1.0.0 release commit, for three unrelated
reasons. None of them affect the numerics.

- The rustdoc/docs.rs job ran `cargo rustdoc -p <crate> --all-features -- --cfg
  docsrs`. Under `--all-features` both crates expose several buildable targets
  (the `exp0` bin, examples, benches), and cargo refuses to forward trailing
  rustdoc arguments to more than one target, so the step could never pass. Pass
  `--lib`. The `cargo doc` form documented in AGENTS.md does not have this
  problem, which is why the breakage went unnoticed.

- The repository had no `.gitattributes`, so Windows checkouts rewrote LF to
  CRLF inside the JSON/JSONL test fixtures. Those fixtures are hashed
  byte-for-byte and compared against committed SHA-256 digests, so the digests
  stopped matching and the pid-core and pid-runlog golden-fixture tests failed
  on Windows only. Pin line endings to LF and mark the byte-hashed assets
  (test fixtures, fuzz corpus) `-text` so git never translates them.

- test_categorical_encoding_is_invariant_to_label_order_and_magnitude passed
  `np.where(...)` output directly to the bindings. With Python int scalars that
  yields the platform default integer dtype, which is int32 on Windows under
  NumPy 1.x, while the bindings take int64; rust-numpy then reports the failed
  downcast as "'ndarray' object is not an instance of 'ndarray'". Pin the dtype,
  matching the `.astype(np.int64)` already used by the neighbouring test.

+++ crates/pid-python/tests/test_v1.py
@@ -113,10 +113,12 @@ def test_signed_labels_noncontiguous_and_read_only_are_safe():
 def test_categorical_encoding_is_invariant_to_label_order_and_magnitude():
     s1, s2, target = gate2([(0, 0, 0), (0, 1, 1), (1, 0, 1), (1, 1, 0)])
     baseline = pid.compute_categorical_sxpid2(s1, s2, target)
+    # `np.where` with Python int scalars yields the platform default integer dtype, which is
+    # int32 on Windows under NumPy 1.x; the bindings take int64. Pin the dtype explicitly.
     relabeled = pid.compute_categorical_sxpid2(
-        np.where(s1 == 0, 10_000, -3),
-        np.where(s2 == 0, -9_000, 17),
-        np.where(target == 0, 42, -1_000_000),
+        np.where(s1 == 0, 10_000, -3).astype(np.int64),
+        np.where(s2 == 0, -9_000, 17).astype(np.int64),
+        np.where(target == 0, 42, -1_000_000).astype(np.int64),
     )
     assert tuple(
         atom.net_nats
